requirements.txt
ansicolors==1.1.8
click==7.1.2
Cython==0.29.21
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Werkzeug==1.0.1
